requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

# The following packages are required for the Arc Invt function app
azure-identity
azure-mgmt-resourcegraph
azure-monitor-query
pandas
